+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **General Changes**
   - Alerts added to warn a user before a timeout to give a chance to stay active
   - Improved visibility of notifications bell icon
+  - Started work on a new landing page
 - **Data Quality Checks**
   - Project editors can now define data quality checks that are used to identify
     sheets that have inconsistent data